+++ decidim-admin/spec/commands/decidim/admin/update_component_spec.rb
@@ -23,20 +23,20 @@ module Decidim::Admin
         settings: {
           dummy_global_attribute_1: true,
           dummy_global_attribute_2: false,
-          disabled_attribute: false
+          readonly_attribute: false
         },
         default_step_settings: {
           step.id.to_s => {
             dummy_step_attribute_1: true,
             dummy_step_attribute_2: false,
-            disabled_step_attribute: false
+            readonly_step_attribute: false
           }
         },
         step_settings: {
           step.id.to_s => {
             dummy_step_attribute_1: true,
             dummy_step_attribute_2: false,
-            disabled_step_attribute: false
+            readonly_step_attribute: false
           }
         }
       )
@@ -45,7 +45,7 @@ module Decidim::Admin
     describe "when valid" do
       let(:valid) { true }
 
-      it "broadcasts :ok and updates the component (except the disable attribute)" do
+      it "broadcasts :ok and updates the component (except the readonly attribute)" do
         expect do
           described_class.call(form, component)
         end.to broadcast(:ok)
@@ -54,12 +54,12 @@ module Decidim::Admin
         expect(component.weight).to eq(3)
         expect(component.settings.dummy_global_attribute_1).to eq(true)
         expect(component.settings.dummy_global_attribute_2).to eq(false)
-        expect(component.settings.disabled_attribute).to eq(true)
+        expect(component.settings.readonly_attribute).to eq(true)
 
         step_settings = component.step_settings[step.id.to_s]
         expect(step_settings.dummy_step_attribute_1).to eq(true)
         expect(step_settings.dummy_step_attribute_2).to eq(false)
-        expect(step_settings.disabled_step_attribute).to eq(true)
+        expect(step_settings.readonly_step_attribute).to eq(true)
       end
 
       it "fires the hooks" do