+++ decidim-core/lib/decidim/settings_manifest.rb
@@ -105,7 +105,7 @@ module Decidim
       attribute :editor, Boolean, default: false
       attribute :required, Boolean, default: false
       attribute :required_for_authorization, Boolean, default: false
-      attribute :disabled
+      attribute :readonly
       attribute :choices
 
       validates :type, inclusion: { in: TYPES.keys }
@@ -122,8 +122,8 @@ module Decidim
         choices.try(:call) || choices
       end
 
-      def disabled?(context)
-        disabled&.call(context)
+      def readonly?(context)
+        readonly&.call(context)
       end
     end
   end