+++ decidim-proposals/lib/decidim/proposals/component.rb
@@ -39,7 +39,7 @@ Decidim.register_component(:proposals) do |component|
     settings.attribute :collaborative_drafts_enabled, type: :boolean, default: false
     settings.attribute :participatory_texts_enabled,
                        type: :boolean, default: false,
-                       disabled: ->(context) { Decidim::Proposals::Proposal.where(component: context[:component]).any? }
+                       readonly: ->(context) { Decidim::Proposals::Proposal.where(component: context[:component]).any? }
     settings.attribute :amendments_enabled, type: :boolean, default: false
     settings.attribute :amendments_wizard_help_text, type: :text, translated: true, editor: true, required: false
     settings.attribute :announcement, type: :text, translated: true, editor: true