+++ decidim-proposals/spec/lib/decidim/proposals/component_spec.rb
@@ -144,7 +144,7 @@ describe "Proposals component" do # rubocop:disable RSpec/DescribeClass
 
       context "when there are no proposals for the component" do
         it "allows to check the setting" do
-          expect(participatory_texts_enabled_container[:class]).not_to include("disabled")
+          expect(participatory_texts_enabled_container[:class]).not_to include("readonly")
           expect(page).not_to have_content("Cannot interact with this setting if there are existing proposals. Please, create a new `Proposals component` if you want to enable this feature or discard all imported proposals in the `Participatory Texts` menu if you want to disable it.")
         end
 
@@ -164,7 +164,7 @@ describe "Proposals component" do # rubocop:disable RSpec/DescribeClass
         end
 
         it "does NOT allow to check the setting" do
-          expect(participatory_texts_enabled_container[:class]).to include("disabled")
+          expect(participatory_texts_enabled_container[:class]).to include("readonly")
           expect(page).to have_content("Cannot interact with this setting if there are existing proposals. Please, create a new `Proposals component` if you want to enable this feature or discard all imported proposals in the `Participatory Texts` menu if you want to disable it.")
         end