+++ CHANGELOG.md
@@ -57,7 +57,7 @@ Due to [\#5553](https://github.com/decidim/decidim/pull/5553), SSL is turned on 
 - **decidim-budgets**: Projects filter by multiple categories [/#5992](https://github.com/decidim/decidim/pull/5992)
 - **decidim-initiatives**: Add pages for versioning. [\#5935](https://github.com/decidim/decidim/pull/5935)
 - **decidim-core**: Added support for enum settings for components [#6001](https://github.com/decidim/decidim/pull/6001)
-- **decidim-core**: Added support for disabling settings for components [#6001](https://github.com/decidim/decidim/pull/6001)
+- **decidim-core**: Added support for readonly settings for components [#6001](https://github.com/decidim/decidim/pull/6001)
 
 ### Changed